airalarm UI refactor
